// service/src/interfaces/avatax.config.interface.ts
export interface AvataxMerchantConfig {
  accountNumber: string;
  licenseKey: string;
  companyCode: string;
  env: boolean;
  logLevel: string;
  addressValidation: boolean;
  disableDocRec: boolean;
  taxCalculation: string;
  enableLogging: boolean;
  line1: string;
  line2: string;
  line3: string;
  postalCode: string;
  city: string;
  region: string;
  country: string;
}